Refactored the package
