+++ app/src/main/java/com/blaster/updater/ExportUpdateService.java
@@ -1,5 +1,6 @@
 /*
  * Copyright (C) 2017-2022 The LineageOS Project
+ * Copyright (C) PixelBlaster-OS
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.lineageos.updater;
+package com.blaster.updater;
 
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
@@ -27,8 +28,9 @@ import android.util.Log;
 import android.widget.Toast;
 
 import androidx.core.app.NotificationCompat;
+import com.blaster.updater.R;
 
-import org.lineageos.updater.misc.FileUtils;
+import com.blaster.updater.misc.FileUtils;
 
 import java.io.File;
 import java.io.IOException;
