+++ app/src/main/java/com/blaster/updater/UpdatesCheckReceiver.java
@@ -1,5 +1,6 @@
 /*
  * Copyright (C) 2017-2022 The LineageOS Project
+ * Copyright (C) PixelBlaster-OS
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.lineageos.updater;
+package com.blaster.updater;
 
 import android.app.AlarmManager;
 import android.app.NotificationChannel;
@@ -29,10 +30,12 @@ import android.util.Log;
 import androidx.core.app.NotificationCompat;
 import androidx.preference.PreferenceManager;
 
+import com.blaster.updater.download.DownloadClient;
+import com.blaster.updater.misc.Constants;
+import com.blaster.updater.misc.Utils;
+
 import org.json.JSONException;
-import org.lineageos.updater.download.DownloadClient;
-import org.lineageos.updater.misc.Constants;
-import org.lineageos.updater.misc.Utils;
+import com.blaster.updater.R;
 
 import java.io.File;
 import java.io.IOException;
