+++ app/src/main/java/com/blaster/updater/UpdatesListActivity.java
@@ -1,5 +1,6 @@
 /*
  * Copyright (C) 2017 The LineageOS Project
+ * Copyright (C) PixelBlaster-OS
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,11 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.lineageos.updater;
+package com.blaster.updater;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import org.lineageos.updater.model.UpdateInfo;
+import com.blaster.updater.model.UpdateInfo;
 
 public abstract class UpdatesListActivity extends AppCompatActivity {
     public abstract void exportUpdate(UpdateInfo update);
