+++ app/src/main/java/com/blaster/updater/download/HttpURLConnectionClient.java
@@ -1,5 +1,6 @@
 /*
  * Copyright (C) 2017-2022 The LineageOS Project
+ * Copyright (C) PixelBlaster-OS
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.lineageos.updater.download;
+package com.blaster.updater.download;
 
 import android.os.SystemClock;
 import android.util.Log;
