+++ app/src/main/java/com/blaster/updater/misc/StringGenerator.java
@@ -1,5 +1,6 @@
 /*
  * Copyright (C) 2017-2022 The LineageOS Project
+ * Copyright (C) PixelBlaster-OS
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,12 +14,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.lineageos.updater.misc;
+package com.blaster.updater.misc;
 
 import android.content.Context;
 import android.content.res.Resources;
 
-import org.lineageos.updater.R;
+import com.blaster.updater.R;
 
 import java.text.DateFormat;
 import java.util.Date;
