+++ app/src/main/java/com/blaster/updater/misc/Utils.java
@@ -1,5 +1,6 @@
 /*
  * Copyright (C) 2017-2023 The LineageOS Project
+ * Copyright (C) PixelBlaster-OS
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.lineageos.updater.misc;
+package com.blaster.updater.misc;
 
 import android.app.AlarmManager;
 import android.content.ClipData;
@@ -32,15 +33,16 @@ import android.widget.Toast;
 
 import androidx.preference.PreferenceManager;
 
+import com.blaster.updater.controller.UpdaterService;
+import com.blaster.updater.model.Update;
+import com.blaster.updater.model.UpdateBaseInfo;
+import com.blaster.updater.model.UpdateInfo;
+
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.lineageos.updater.R;
-import org.lineageos.updater.UpdatesDbHelper;
-import org.lineageos.updater.controller.UpdaterService;
-import org.lineageos.updater.model.Update;
-import org.lineageos.updater.model.UpdateBaseInfo;
-import org.lineageos.updater.model.UpdateInfo;
+import com.blaster.updater.R;
+import com.blaster.updater.UpdatesDbHelper;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -50,7 +52,6 @@ import java.util.ArrayList;
 import java.util.Enumeration;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Locale;
 import java.util.Set;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
@@ -382,7 +383,6 @@ public class Utils {
         return preferences.getInt(Constants.PREF_AUTO_UPDATES_CHECK_INTERVAL,
                 Constants.AUTO_UPDATES_CHECK_INTERVAL_WEEKLY);
     }
-
     public static boolean isUpdateCheckEnabled(Context context) {
         return getUpdateCheckSetting(context) != Constants.AUTO_UPDATES_CHECK_INTERVAL_NEVER;
     }
